Remove duplicate instance of SmilHelper
