+++ QKSMS/src/main/java/com/moez/QKSMS/model/AudioModel.java
@@ -28,7 +28,7 @@ import android.util.Log;
 import com.android.mms.dom.events.EventImpl;
 import com.android.mms.dom.smil.SmilMediaElementImpl;
 import com.google.android.mms.MmsException;
-import com.google.android.mms.smil.SmilHelper;
+import com.moez.QKSMS.common.SmilHelper;
 import com.moez.QKSMS.common.SqliteWrapper;
 import com.moez.QKSMS.exception.ContentRestrictionException;
 import org.w3c.dom.events.Event;
