+++ QKSMS/src/main/java/com/moez/QKSMS/model/VideoModel.java
@@ -30,9 +30,9 @@ import com.android.mms.dom.events.EventImpl;
 import com.android.mms.dom.smil.SmilMediaElementImpl;
 import com.google.android.mms.ContentType;
 import com.google.android.mms.MmsException;
-import com.google.android.mms.smil.SmilHelper;
 import com.moez.QKSMS.QKSMSApp;
 import com.moez.QKSMS.common.LogTag;
+import com.moez.QKSMS.common.SmilHelper;
 import com.moez.QKSMS.common.SqliteWrapper;
 import com.moez.QKSMS.common.google.ItemLoadedCallback;
 import com.moez.QKSMS.common.google.ItemLoadedFuture;
